+++ tests/Classes/User.php
@@ -2,10 +2,12 @@
 
 namespace Luma\Tests\Classes;
 
+use Luma\AuroraDatabase\Attributes\AuroraCollection;
 use Luma\AuroraDatabase\Attributes\Column;
 use Luma\AuroraDatabase\Attributes\Identifier;
 use Luma\AuroraDatabase\Attributes\Schema;
 use Luma\AuroraDatabase\Model\Aurora;
+use Luma\AuroraDatabase\Utils\Collection;
 
 #[Schema('DatabaseComponentTest')]
 class User extends Aurora
@@ -19,6 +21,15 @@ class User extends Aurora
 
     #[Column('strEmailAddress')]
     private string $strEmailAddress;
+
+    #[Column('strPassword')]
+    private string $password;
+
+    #[Column('dtmCreated')]
+    private \DateTimeInterface $created;
+
+    #[AuroraCollection(class: Article::class, property: 'author')]
+    private Collection $articles;
 
     /**
      * @return string
@@ -34,5 +45,13 @@ class User extends Aurora
     public function getEmailAddress(): string
     {
         return $this->strEmailAddress;
+    }
+
+    /**
+     * @return Collection<Article>
+     */
+    public function getArticles(): Collection
+    {
+        return $this->articles;
     }
 }