+++ CHANGELOG.md
@@ -5,9 +5,33 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres 
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - 2024-03-24
+### Added
+- Added new `Collection` class
+- Added new `AuroraCollection` attribute
+- Added basic handling for OneToMany by way of the `AuroraCollection` attribute and appropriate property setup
+- Can now call the `->with([AssociatedClass::class])` method to fetch associated models
+
+### Changed
+- N/A
+
+### Deprecated
+- N/A
+
+### Removed
+- N/A
+
+### Fixed
+- N/A
+
+### Security
+- N/A
+
+---
+
 ## [2.1.0] - 2024-03-22
 ### Added
-- Added additional argument to `AuroraMapper::map` to take an instance of the class to map into.
+- Added additional argument to `AuroraMapper::map` to take the FQCN of the class to map into.
 
 ### Changed
 - Renamed `Aurora::make` -> `Aurora::create`